Move api calls from statsboard to Datacontext
