+++ src/contexts/DataContext.jsx
@@ -8,6 +8,10 @@ export const DataProvider = ({ children }) => {
   const [countryData, setCountryData] = useState(null);
   const [vaccineData, setVaccineData] = useState(null);
   const [countryCSVData, setCountryCSVData] = useState(null);
+  const [cases, setCases] = useState("");
+  const [casesToday, setCasesToday] = useState("");
+  const [deaths, setDeaths] = useState("");
+  const [deathsToday, setDeathsToday] = useState("");
 
   // functions
   const checkContinents = () => {
@@ -48,6 +52,46 @@ export const DataProvider = ({ children }) => {
       });
   };
 
+  const fetchTotalCases = () => {
+    return fetch("https://disease.sh/v3/covid-19/all")
+      .then((response) => {
+        return response.json();
+      })
+      .then((data) => {
+        return setCases(data.cases);
+      });
+  };
+
+  const fetchTotalCasesToday = () => {
+    return fetch("https://disease.sh/v3/covid-19/all")
+      .then((response) => {
+        return response.json();
+      })
+      .then((data) => {
+        return setCasesToday(data.todayCases);
+      });
+  };
+
+  const fetchTotalDeaths = () => {
+    return fetch("https://disease.sh/v3/covid-19/all")
+      .then((response) => {
+        return response.json();
+      })
+      .then((data) => {
+        return setDeaths(data.deaths);
+      });
+  };
+
+  const fetchTotalDeathsToday = () => {
+    return fetch("https://disease.sh/v3/covid-19/all")
+      .then((response) => {
+        return response.json();
+      })
+      .then((data) => {
+        return setDeathsToday(data.todayDeaths);
+      });
+  };
+
   return (
     <DataContext.Provider
       value={{
@@ -62,6 +106,14 @@ export const DataProvider = ({ children }) => {
         checkContinents,
         checkCountries,
         checkVaccines,
+        fetchTotalCases,
+        fetchTotalCasesToday,
+        fetchTotalDeaths,
+        fetchTotalDeathsToday,
+        cases,
+        casesToday,
+        deaths,
+        deathsToday,
       }}
     >
       {children}
